Set minimum required version of Perl to 5.028
The minimum required version of Perl was stated in the cpanfile
to be `1.028`, but this in all likelihood was a mistake, since
both modules in this distribution do `use v5.28`.

This change updates the cpanfile so it reports a more meaningful
minimum version of Perl.

Note that [perlver] reports that the minimum version from a syntax
point of view is actually 5.20, but that seemed like too big a
change.

[perlver]: https://metacpan.org/pod/Perl::MinimumVersion